moved database as only used by BufferManager
