+++ spinn_front_end_common/interface/buffer_management/__init__.py
@@ -1,3 +1,6 @@
-from .buffer_manager import BufferManager
 
-__all__ = ["BufferManager"]
+from .abstract_database import AbstractDatabase
+from .buffer_manager import BufferManager
+from .sqllite_database import SqlLiteDatabase
+
+__all__ = ["AbstractDatabase", "BufferManager", "SqlLiteDatabase"]
