+++ spinn_front_end_common/interface/buffer_management/db.sql
