--- spinn_front_end_common/interface/database/__init__.py
@@ -1,4 +1,0 @@
-from .abstract_database import AbstractDatabase
-from .sqllite_database import SqlLiteDatabase
-
-__all__ = ["AbstractDatabase", "SqlLiteDatabase"]
